Removed references to STL and migrated target code to C++17.

+++ software/code/arduino/src/clock/clock.hpp
@@ -1,8 +1,7 @@
 #ifndef CLOCK_H
 #define CLOCK_H
 
-#include <cstdint>
 
-typedef int32_t (*GetCurrentTimeSecondsCallback)();
+typedef long (*GetCurrentTimeSecondsCallback)();
 
 #endif
