+++ software/code/arduino/src/system_timer/system_timer.hpp
@@ -2,10 +2,9 @@
 #define SYSTEM_TIMER_H
 
 #include "clock/clock.hpp"
-#include <cstdint>
 
 void SystemTimer_Initialize();
-int32_t SystemTimer_GetCurrentTimeSeconds();
-int32_t SystemTimer_GetCurrentTimeMinutes();
+long SystemTimer_GetCurrentTimeSeconds();
+long SystemTimer_GetCurrentTimeMinutes();
 
 #endif
