+++ software/tests/core/mocks/src/wrap_up_action_mock.cpp
@@ -15,7 +15,7 @@ void WrapUpActionMock::WrapUp()
     this->number_of_calls++;
 }
 
-int32_t WrapUpActionMock::GetNumberOfCalls()
+long WrapUpActionMock::GetNumberOfCalls()
 {
     return this->number_of_calls;
 }
